requirements.txt
asgiref==3.4.1
Django==3.2.6
inflect==5.3.0
Pillow==8.3.1
python-magic==0.4.24
python-magic-bin==0.4.14
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
